Rename RadioControl with ButtonGroupControl
