+++ src/components/arrow-control/index.js
@@ -7,7 +7,7 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import AdvancedNumberControl from '../advanced-number-control';
-import RadioControl from '../radio-control';
+import ButtonGroupControl from '../button-group-control';
 import ToggleSwitch from '../toggle-switch';
 import InfoBox from '../info-box';
 import {
@@ -101,7 +101,7 @@ const ArrowControl = props => {
 			/>
 			{props['arrow-status'] && (
 				<>
-					<RadioControl
+					<ButtonGroupControl
 						label=''
 						selected={getLastBreakpointAttribute(
 							'arrow-side',
