+++ src/components/index.js
@@ -5,9 +5,9 @@ export { default as ArrowControl } from './arrow-control';
 export { default as ArrowDisplayer } from './arrow-displayer';
 export { default as AxisControl } from './axis-control';
 export { default as BackgroundControl } from './background-control';
-export { default as BlockBackgroundControl } from './background-control/blockBackgroundControl';
 export { default as BackgroundDisplayer } from './background-displayer';
 export { default as BaseControl } from './base-control';
+export { default as BlockBackgroundControl } from './background-control/blockBackgroundControl';
 export { default as BlockPlaceholder } from './block-placeholder';
 export { default as BlockResizer } from './block-resizer';
 export { default as BlockStylesControl } from './block-styles-control';
@@ -15,6 +15,7 @@ export { default as BorderControl } from './border-control';
 export { default as BoxShadowControl } from './box-shadow-control';
 export { default as Breadcrumbs } from './breadcrumbs';
 export { default as Button } from './button';
+export { default as ButtonGroupControl } from './button-group-control';
 export { default as CheckBoxControl } from './checkbox-control';
 export { default as ClipPath } from './clip-path-control';
 export { default as ColorControl } from './color-control';
@@ -46,7 +47,6 @@ export { default as OpacityControl } from './opacity-control';
 export { default as OverflowControl } from './overflow-control';
 export { default as Placeholder } from './placeholder';
 export { default as PositionControl } from './position-control';
-export { default as RadioControl } from './radio-control';
 export { default as RawHTML } from './raw-html';
 export { default as ResponsiveControl } from './responsive-control';
 export { default as ResponsiveTabsControl } from './responsive-tabs-control';
