+++ src/components/toolbar/components/background-color/index.js
@@ -8,7 +8,7 @@ import { __ } from '@wordpress/i18n';
  */
 import ToolbarPopover from '../toolbar-popover';
 import ColorLayer from '../../../background-control/colorLayer';
-import RadioControl from '../../../radio-control';
+import ButtonGroupControl from '../../../button-group-control';
 import {
 	getBlockStyle,
 	getColorRGBAString,
@@ -102,7 +102,7 @@ const BackgroundColor = props => {
 			icon={<div className='toolbar-item__icon' style={getStyle()} />}
 		>
 			<div className='toolbar-item__background__popover'>
-				<RadioControl
+				<ButtonGroupControl
 					label={__('Enable Background Colour', 'maxi-blocks')}
 					selected={isBackgroundColor}
 					options={[
