+++ src/components/toolbar/components/icon-background/index.js
@@ -8,7 +8,7 @@ import { __ } from '@wordpress/i18n';
  */
 import ToolbarPopover from '../toolbar-popover';
 import ColorControl from '../../../color-control';
-import RadioControl from '../../../radio-control';
+import ButtonGroupControl from '../../../button-group-control';
 import {
 	getColorRGBAString,
 	getDefaultAttribute,
@@ -53,7 +53,7 @@ const IconBackground = props => {
 			advancedOptions='icon'
 		>
 			<div className='toolbar-item__icon-background__popover'>
-				<RadioControl
+				<ButtonGroupControl
 					label={__(
 						'Inherit Colour/Backgrond from Button',
 						'maxi-block'
